source/emol/requirements.txt
Authomatic==0.1.0.post1
Flask==0.11.1
Flask-Login==0.4.0
Flask-RESTful==0.3.5
Flask-SQLAlchemy==2.2
Jinja2==2.8
MarkupSafe==0.23
PyMySQL==0.7.11
SQLAlchemy==1.1.9
Unidecode==0.04.19
Werkzeug==0.11.11
aniso8601==1.2.0
astroid==1.4.9
blinker==1.4
click==6.6
flup6==1.1.1
isort==4.2.5
itsdangerous==0.24
lazy-object-proxy==1.2.2
mccabe==0.6.1
pep257==0.7.0
py==1.4.33
pyaes==1.6.0
pylint==1.6.5
pylint-flask==0.5
pylint-plugin-utils==0.2.4
pytest==3.0.7
pytest-flask==0.10.0
python-dateutil==2.6.0
python-slugify==1.2.1
pytz==2016.7
requests==2.12.3
six==1.10.0
wrapt==1.10.8
